[Ingest Manager] Do not allow empty namespace (#67492)
