+++ x-pack/plugins/ingest_manager/server/types/models/agent_config.ts
@@ -9,7 +9,7 @@ import { AgentConfigStatus } from '../../../common';
 
 const AgentConfigBaseSchema = {
   name: schema.string(),
-  namespace: schema.maybe(schema.string()),
+  namespace: schema.string({ minLength: 1 }),
   description: schema.maybe(schema.string()),
   monitoring_enabled: schema.maybe(
     schema.arrayOf(schema.oneOf([schema.literal('logs'), schema.literal('metrics')]))
