+++ x-pack/test/api_integration/apis/index.js
@@ -27,7 +27,7 @@ export default function ({ loadTestFile }) {
     loadTestFile(require.resolve('./short_urls'));
     loadTestFile(require.resolve('./lens'));
     loadTestFile(require.resolve('./fleet'));
-    loadTestFile(require.resolve('./ingest'));
+    loadTestFile(require.resolve('./ingest_manager'));
     loadTestFile(require.resolve('./endpoint'));
     loadTestFile(require.resolve('./ml'));
   });
